Fix iclear code for lua 5.0 (mods not applying)
NotITG's Lua 5.0 is really cursed in that the # operator isn't something
that you can rely on. The # operator is implemented by calling
table.getn, which is a gullible function that doesn't play nicely with
all table modifications. My understanding is that after calling
table.insert or table.remove on a table, the table module caches the
length and refuses to re-verify it if the table was modified in any way
other than through the table module. This understanding could be wrong,
but it seems that avoiding table.remove in this situation fixes the
issue.